Split admin event log entry presentation
