+++ app/components/admin/event_log_entry_component.html.erb
@@ -8,9 +8,7 @@
       <div class="truncate text-sm text-slate-600">
         <%= render EventDescriptionComponent.new(event: event) %>
       </div>
-      <% if (label = subject_label) %>
-        <div class="text-xs text-slate-500"><%= label %></div>
-      <% end %>
+      <div class="text-xs text-slate-500"><%= safe_join([user_label, subject_label].compact, " • ") %></div>
     </div>
     <%= link_to helpers.short_time_ago(event.created_at), href,
                 class: "shrink-0 text-xs font-medium text-slate-500 hover:text-slate-700",
