+++ app/views/admin/events/index.html.erb
@@ -1,7 +1,7 @@
-<% content_for :title, "Events" %>
+<% content_for :title, "System Events" %>
 
 <div class="space-y-8">
-  <%= render PageHeaderComponent.new(title: "Events") do |header| %>
+  <%= render PageHeaderComponent.new(title: "System Events") do |header| %>
     <% if @filter.present? %>
       <% header.with_context_paragraph do %>
         <span class="text-sm text-slate-500" data-key="admin.events.filter-summary">
